Made SQS_DELETE_ENABLED configurable and defaults to false

+++ analytics/backend_sqs.go
@@ -130,48 +130,56 @@ func (q *SQSReaderImpl) GetMessages(waitTimeout int64, maxNumberOfMessages int64
 
 // DeleteMessages deletes a single message from SQS using the message receipt handle
 func (q *SQSReaderImpl) DeleteMessage(receiptHandle string) (*sqs.DeleteMessageOutput, error) {
-	params := sqs.DeleteMessageInput{
-		QueueUrl:      aws.String(q.URL),
-		ReceiptHandle: aws.String(receiptHandle),
+	if config.SQSDeleteEnabled {
+		params := sqs.DeleteMessageInput{
+			QueueUrl:      aws.String(q.URL),
+			ReceiptHandle: aws.String(receiptHandle),
+		}
+
+		// Generate the request
+		req := q.Client.DeleteMessageRequest(&params)
+		// Send the request
+		resp, err := req.Send()
+
+		if err != nil {
+			return nil, fmt.Errorf("failed to delete message with receipt %s, %v", receiptHandle, err)
+		}
+
+		return resp, err
+	} else {
+		return nil, fmt.Errorf("deletion of SQS messages is disabled")
 	}
-
-	// Generate the request
-	req := q.Client.DeleteMessageRequest(&params)
-	// Send the request
-	resp, err := req.Send()
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to delete message with receipt %s, %v", receiptHandle, err)
-	}
-
-	return resp, err
 }
 
 // BatchDeleteMessages deletes one or many message(s) from SQS using the message receipt handle(s)
 func (q *SQSReaderImpl) BatchDeleteMessages(receiptHandles []string) (*sqs.DeleteMessageBatchOutput, error) {
-	entries := make([]sqs.DeleteMessageBatchRequestEntry, len(receiptHandles))
+	if config.SQSDeleteEnabled {
+		entries := make([]sqs.DeleteMessageBatchRequestEntry, len(receiptHandles))
 
-	// Generate list of sqs.DeleteMessageBatchRequestEntry using the message receipt handle(s)
-	for i, receiptHandle := range receiptHandles {
-		entries[i] = sqs.DeleteMessageBatchRequestEntry{
-			Id:            aws.String(strconv.Itoa(i)),
-			ReceiptHandle: aws.String(receiptHandle),
+		// Generate list of sqs.DeleteMessageBatchRequestEntry using the message receipt handle(s)
+		for i, receiptHandle := range receiptHandles {
+			entries[i] = sqs.DeleteMessageBatchRequestEntry{
+				Id:            aws.String(strconv.Itoa(i)),
+				ReceiptHandle: aws.String(receiptHandle),
+			}
 		}
+
+		params := sqs.DeleteMessageBatchInput{
+			Entries:  entries,
+			QueueUrl: aws.String(q.URL),
+		}
+
+		// DeleteMessages deletes a single message from SQS using the message receipt handle
+		req := q.Client.DeleteMessageBatchRequest(&params)
+		// Send the request
+		resp, err := req.Send()
+
+		if err != nil {
+			return nil, fmt.Errorf("failed to batch delete messages with receipt, %v", err)
+		}
+
+		return resp, err
+	} else {
+		return nil, fmt.Errorf("deletion of SQS messages is disabled")
 	}
-
-	params := sqs.DeleteMessageBatchInput{
-		Entries:  entries,
-		QueueUrl: aws.String(q.URL),
-	}
-
-	// DeleteMessages deletes a single message from SQS using the message receipt handle
-	req := q.Client.DeleteMessageBatchRequest(&params)
-	// Send the request
-	resp, err := req.Send()
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to batch delete messages with receipt, %v", err)
-	}
-
-	return resp, err
 }
